initial move for systemsManagerSetParameter
